src/test/java/ru/job4j/condition/MaxTest.java
package ru.job4j.condition;

import org.junit.jupiter.api.Test;
import static org.assertj.core.api.Assertions.assertThat;

public class MaxTest {
    @Test
    public void whenMax1To2Then2() {
        int left = 1;
        int right = 2;
        int result = Max.getMax(left, right);
        int expected = 2;
        assertThat(result).isEqualTo(expected);
    }

    @Test
    public void whenMax2To1Then2() {
        int left = 2;
        int right = 1;
        int result = Max.getMax(left, right);
        int expected = 2;
        assertThat(result).isEqualTo(expected);
    }

    @Test
    public void whenMax4To4Then4() {
        int left = 4;
        int right = 4;
        int result = Max.getMax(left, right);
        int expected = 4;
        assertThat(result).isEqualTo(expected);
    }

    @Test
    public void whenMax4To5To6Then6() {
        int first = 4;
        int second = 5;
        int third = 6;
        int result = Max.getMax(first, second, third);
        int expected = 6;
        assertThat(result).isEqualTo(expected);
    }

    @Test
    public void whenMax4To5To6To7Then6() {
        int first = 4;
        int second = 5;
        int third = 6;
        int fourth = 7;
        int result = Max.getMax(first, second, third, fourth);
        int expected = 7;
        assertThat(result).isEqualTo(expected);
    }
}